[SES-5085] - Pro badge not showing for pro who accepting a MR
