+++ app/src/main/java/org/session/libsession/messaging/messages/Message.kt
@@ -1,5 +1,8 @@
 package org.session.libsession.messaging.messages
 
+import network.loki.messenger.libsession_util.protocol.ProFeature
+import network.loki.messenger.libsession_util.protocol.ProMessageFeature
+import network.loki.messenger.libsession_util.protocol.ProProfileFeature
 import network.loki.messenger.libsession_util.util.ExpiryMode
 import org.session.libsession.database.MessageDataProvider
 import org.session.libsession.messaging.MessagingModuleConfiguration
@@ -10,6 +13,8 @@ import org.session.libsession.utilities.Address
 import org.session.protos.SessionProtos
 import org.session.protos.SessionProtos.Content.ExpirationType
 import org.thoughtcrime.securesms.database.model.MessageId
+import org.thoughtcrime.securesms.pro.toProMessageBitSetValue
+import org.thoughtcrime.securesms.pro.toProProfileBitSetValue
 
 abstract class Message {
     var id: MessageId? = null // Message ID in the database. Not all messages will be saved to db.
@@ -26,6 +31,20 @@ abstract class Message {
     var specifiedTtl: Long? = null
 
     var expiryMode: ExpiryMode = ExpiryMode.NONE
+
+    /**
+     * The pro features enabled for this message.
+     *
+     * Note:
+     * * When this message is an incoming message, the pro features will only be populated
+     * if we can prove that the sender has an active pro subscription.
+     *
+     * * When this message represents an outgoing message, this property can be populated by
+     * application code at their wishes but the actual translating to protobuf onto the wired will
+     * be checked against the current user's pro proof, if no active pro subscription is found,
+     * the pro features will not be sent in the protobuf messages.
+     */
+    var proFeatures: Set<ProFeature> = emptySet()
 
     open val coerceDisappearAfterSendToRead = false
 
@@ -68,6 +87,19 @@ abstract class Message {
 
         // * Timestamps
         builder.setSigTimestamp(sentTimestamp!!)
+
+        // Pro features
+        if (proFeatures.any { it is ProMessageFeature }) {
+            builder.proMessageBuilder.setMsgBitset(
+                proFeatures.toProMessageBitSetValue()
+            )
+        }
+
+        if (proFeatures.any { it is ProProfileFeature }) {
+            builder.proMessageBuilder.setProfileBitset(
+                proFeatures.toProProfileBitSetValue()
+            )
+        }
 
         // Then ask the subclasses to build their specific proto
         buildProto(builder, messageDataProvider)
